[functional tests] implement address substitution
This implements simple text substitution in functional tests, allowing addresses to be filled in.
```
//! account: bob, 1000000

main() {
    ...
    addr = {{bob}}
    // will be replaced by bob's address
    ...
}
```
~~Note: this is a temporary workaround which will get replaced by transaction arguments once those are supported.
However that will take a bit longer to come, along with support for multiple transactions. This is simple to implement so I did it, to enable some of the disabled payment tests first.~~
I'm taking my words back. We need this to be able to refer to modules published in previous transactions once support for running multiple transactions is added.