Revert "[network] limit number of connection attempts to disconnected peer"
This reverts commit https://github.com/libra/libra/commit/2179950e867c9b269c9d8ca92d2f126bf19807ed.
The above commit introduced a limit on connection attempts from 1 peer
to another. A side-effect of this behavior is that after a period of
network congestion we could have all nodes disconnected from each other,
and no longer attempting to establish connections, having exhausted the
attempts. A more robust solution is to have an exponential back-off for
each peer, but adjust that by the # of disconnected peers. This would
ensure that we are disconnected from a large % of our peers, we would
keep the retry window low.