[consensus] Make consensus data cryptographically agile
*What this does*:
This PR offers a binding discipline for using cryptographic types in a pseudo-generic way in the consensus crate, without propagating generic parameters. Instead of making consensus structures parametric in their cryptographic type (as in #522), we make consensus data structures depend on instantiations of generic structs.
- those instantiations are all aliased in one central place, allwoing an easy change of the cryptographic scheme.
- they are defined *in terms of a generic struct*, which, by virtue of genericity ensures encapsulation.

This PR also introduces a newtype pattern for the signature type sued in `BlockSource`, `VoteMsg`, `TiemOutMsg`, etc ... It is built as an alternative to #522 that offers a better readability through lack of generic type parameters.

This relies on practices that will have to be enforced by convention in the consensus crate:
1. nobody uses cryptographic schemes directly.
2. when defining a struct that must use cryptographic types, define it in the form of a polymorphic container struct that works for any `impl` of some `crypto::traits` type, then alias that struct type in `consensus::chained_bft::consensus_types`

*Why this is better*:
- Contains less generic parameters than #522,

*Why this is worse*:
- Breaking any one of the (1, 2) practices above will either lose cryptographic agility or encapsulation,

*Test*
- existing unit tests,

Closes #522