[storage] `SchemaBatch` commit content in order
We realized if we put things to a batch out of order they can end up in different memtables, although they are commited in one DB transaction, hence cause unnecessary compactions. We can either solve this by sorting things beforehand, like in #673, or make `SchemaBatch` sort things up before invoking the db batch. This does the latter. This at least addresses the similar issue on transacton/events accumulators, and protect against other situations.

n.b. 1. I do think it's intuitive to think "if things between batches are in order and doesn't overlap, we should not see compactions". 2. WriteBatchWithIndex is not available in the RocksDB binding we are using, but as we usually don't need the read-your-own-write ability (WBWI's main concern), I think it's not too bad a solution to do it in our own wrapper.

Closes: #867
Approved by: zamsden