[executor] use executor directly instead of execution GRPC service
I believe there is no need to make execution a thin and standalone
service. On the contrary, it should be a thin submodule inside
consensus/state_synchronizer for the following reasons:
1. Execution is stateful with a block tree inside. It is not thoughtful to
implement it as a separate service as it require extra work to keep the tree
which has a counterpart in consensus already.
2. Service is usually used for multiple clients to call. But in our case
since the execution is not stateless, we cannot support multiple
callsites simultaneously.
3. GRPC degrades performance compare to direct calling and adds a lot of
complexity in the code. For example, tremendous proto definitions for
requests and responses.

Closes: #939
Approved by: lightmark