[types] Define the block metadata structure so that it can be stored on chain in the future.
This will be our very first step to implement block metadata contract on chain. With this contract, we will be able to get basic info about the consensus from the smart contract side. This PR will implement a struct to represent the info that we want Move to understand about the consensus in Rust. The  next step would be we will pass such struct to VM and write such struct on chain in Move module.

Closes: #1068
Approved by: sblackshear