[enhancement] Refactor libra_channel interface to add an explicit Key
Summary
The current libra_channel interface is agnostic of the concept of a Key. We have pushed the abstraction of Key to ValidatorMessage trait which forces each message to implement a trait which returns AccountAddress

This PR updates the libra_channel interface to accept an arbitrary Key in its put() method which will make it easy on the Sender side to add messages. The Sender side will no longer be forced to implement ValidatorMessage trait.

With this design, we do leak the concept of a Key explicitly to the sender, but we were not hiding it very well with the ValidatorMessage trait either.

Test Plan
Updated unit tests

Closes: #1500
Approved by: ankushagarwal