[Proof] Remove redundant assertion in `new`
We should not have the assertion here. Clients would call `new` when
constructing the proofs from protobuf objects, but the objects returned
by servers are not guaranteed to be correct. So a bad server could cause
the clients to crash.

Since we will verify the proof later, it's probably redundant to do this
particular check in `new`.

Closes: #1715
Approved by: delegate