[config] Add error handling on save
Save historically used expect but that meant that if an application
faile to persist its data, it would crash even if it was a recoverable
event, i.e., persisting data wasn't critical to operations.

Now this returns back a Result.

This affected safety rules substantially and so the API now supports an
"InternalError" rather than crashing upon these events. While it is
up for debate if SafetyRules should crash or just signal further
upstream that this is a blocking error that should retry again later is
delegated to the future.

Closes: #1937
Approved by: davidiw