[language] LibraVM and MoveVm initial separation
This is the initial break up of the Libra execution logic in a LibraVM (name to be finalized) and
a MoveVM.
The separation allows for the 2 layers to move independently and it also cleans up a lot of lifetime
parameters that are not needed any longer.
This also changes the outside type used: it was `MoveVM` and it's changed to `LibraVM`.
The latter will likely change name but I wanted to free the MoveVM name to be used for what it
really is. That is a mechanical change so it should be fine.

The entry points are in move_vm.rs (as it used to) and that module has got more code from other files.
All the entry points into the interpreter have been moved to `VMRuntime` and `TransactionExecutor`
holds much less state than it used to be. `TransactionExecutor` essentially has some
transaction execution logic and all execution is forwarded to `VMRuntime`.

The state machine for the transaction is still around as it used to be, however bytecode verification has
been moved to the execution state. That is in line with the new behavior and something that may evolve still.

There is another clean up that should happen soon and it's the `StateView` and `VMRuntime`
relationship which is something that should be changed by removing the needed `StateView`
and use the data view instead.
In other words, the VM should use the data view (a concept of that) to fetch code.

Overall this code will evolve dramatically in the next 2 weeks so I don't think there is much value
in cleaning it up any more. Evolutions of the Libra layer and the Move layer will likely rewrite a
significant portion of this code base so let's keep "change requests" to a minimum.

Closes: #1946
Approved by: tzakian