Integrates stdlib specs into upstream repo.
This copies the specs for libra coin and account over from the CAV paper repo and adapts them to recent changes to return value notation in post conditions.

Moreover, this sets up the verification tests for those files. A user which has Z3_EXE and BOOGIE_EXE set should be able to run those tests. The tests
currently verify what we have specified for `libra_coin`, but for `libra_account` there is a failure for `withdraw_with_capability`, solution for which
is discussed elsewhere. This one was marked with a TODO and the currently expected diagnosis was marked in the mvir source so it passes as a test for
the remaining part of the `libra_account`.

Closes: #2218
Approved by: cbarrettfb