[language/ir-to-bytecode] use Foo_ for unspanned, Foo for spanned
This follows the convention in Ocaml and in the Move language source compiler.

In the previous scheme, there was an (unspanned) `Var::new_` method which would
produce a (spanned) `Var_`. Now that the situation is reversed, the naming scheme
stopped making sense (`Var_::new_` would produce `Var`) -- so the method has
been removed and its code inlined.

Closes: #2383
Approved by: tnowacki