[move source language] Record the source location of the last semicolon in a Sequence.
One of the tests from https://github.com/libra/libra/pull/2578 revealed a need
to track the source location of the last semicolon in a Sequence, so that
unreachable code errors can report the right location. This adds that information
to the parser AST, without yet using it for the error message.

Closes: #2596
Approved by: tnowacki