[move-prover] Fixing a systematic typo in variable names.
I thought this was fixed already in a previous PR, but I must have looked at a different branch instead of the submitted one.

Also fixes an issues with unstable file names in baselines. If we report an error on generated Boogie, the file name in diagnosis with a temporary unstable location. Added a heuristic to baseline test utility to detect this and remove parts of the unstable file name.

Closes: #3168
Approved by: shazqadeer