[move-prover] Sort of fixing a bytecode verification error which broke Ci.
It appears #3633 broke Ci for the move prover via bytecode verification error in the older copy of the stdlib we are using. The culprit has been identified by inserting debug prints into the bytecode verifier. Because its not obvious why its failing, the according code has been commented out, given that we retire this copy of the stdlib soon anyway.

Closes: #3691
Approved by: DavidLDill