[secure-storage] Merge set and create and remove policy
Almost no libraries had distinct operations for set and create, hence it
wsa alawys awkward to emulate.

Additionally policies aren't prevasive enough to generalize the right
interface, so as long as it is only in vault, we'll have somewhat of a
kludgy means on how to do it right.

For what it is worth, we will not be using policy logic for production
and if we do, we'll need to spend a lot more time working on this
frameowrk.

Closes: #3812
Approved by: joshlind