[crypto] removes the batch feature
The batch feature is not working due to a combination of:
- https://github.com/dalek-cryptography/ed25519-dalek/issues/126 and:
- our x tooling unifying all features, and throwing batch into the mix as soon as it exists, despite it not being used,
- the new reoslver not being switched on (https://github.com/libra/libra/pull/3134)

A fix has been mered upstream in https://github.com/dalek-cryptography/ed25519-dalek/pull/127 but is not released yet, leaving us bereft of a fixed ed25519_dalek version to point to. We could patch the merlin reference on crates.io, except the fixed merlin version is just at the boundary of a major version, and major version incompatibilities halt patch overrides (https://github.com/rust-lang/cargo/issues/5640).

Since there is tooling that urgently relies on correct version resolution to work and is suggesting the rather extreme measure of putting this technical fix in our formally-verified-crypto fork  (https://github.com/novifinancial/ed25519-dalek/pull/8), I think we would rather de-activate the feature entirely. A mention of it in code is explicitly left in (with no side effects) as a reminder to reinstate it once either of the following happens:
- we switch to the new resolver,
- a fixed dalek is released,
- x has the ability to not run with all features.

Closes: #5061
Approved by: rexhoffman