[move-prover] Prover tool installation scripts and CI integration
This PR integrates the installation logic for the Move prover tools in the general libra script at `./scripts/dev-setup.sh`. This script is also used for CI, enabling prover tests to run as part of regular continuous integration testing.

Two new flags are added to `dev-setup.sh`:

- `dev-setup.sh -z` installs or updates only the prover tools.
- `dev-setup.sh -y <other flags` installs the full development environment including the prover tools (and what else is selected with `<other flags>`.

The move-prover documentation as been updated to describe the new installation procedure. With this PR landing, Move prover tests are active in CI.

This PR also adds general support for `dnf` as a package manager in the dev-setup script. This allows f.i. to run the script on a FB devserver. This support is still experimental and has not been deeply tested yet.

Closes: #6213