[stream-ratelimiter] remove stream-ratelimiter crate
From the network fuzzing we discovered a memory leak in the
stream-ratelimiter crate which was incorrectly using the
futures-semaphore crate.

The root problem was using `ManuallyDrop` on the `futures-semaphore`
permits as a work-around for forgetting a permit. Of course, no handles
to the permits were stored and so the permits were never dropped, which
would mean the underlying `Arc<tokio::sync::Semaphore>` would not get
freed since the reference count would be > 0.

Though we have an actual fix, we decided to just remove
stream-ratelimiter altogether in the meantime, since we're already
working on a better replacement token-bucket filter.