[move-prover] Improving bytecode dump output + adding a `modifies` example.
This improves the bytecode dump by marking in the generated code which opaque function calls have been expanded to pre/post conditions, plus a few more improvements. It also adds a further example to `tests/xsources/design` which illustrates how the `modifies` clause is compiled.

The change in bytecode dump leads to many baseline changes for the bytecode pipeline tests.