[network] Use NetworkId instead of RoleType to identify networks
RoleType was representing what kind of node the code is running, and the
type of networks, which was making it very confusing to tell when there
were "full node" networks on validators.  RoleType is now properly set
for the network context, and is not used to make Network decisions, just
for logging.

There was additionally a gap in logic for default network configs used
for tests and the network configs loaded in real situations.  This
ensures all are properly handled the same.

Closes: #7604