[diem framework] Fix some links in script docs, make explanations clearer
There was a number of places where the idea of "authentication key" and
"authentication key prefix" was hard to understand, so added pointers to
documentation for this in each transaction script.

Additionally, clarified language around public/authentication keys in
`SharedEd25519` scripts, and how the public key being passed relates to
the new authentication key for the account.

Also fixed some broken links in the larger script docs.

Closes: #8050