[move-lang] ensure that `dependency_order`  is unique for each module
The code for building a dependency graph does not add nodes for modules with no dependencies. Such modules will be each assigned a `dependency_order` of 0. Each module with dependencies has a unique `dependency_order`, so it's probably good to enforce this invariant for all modules.

This PR achieves this by adding a node to the dep graph for modules with no dependencies.

Closes: #8252