[move-prover] support old(..) in inlined spec blocks
In addition to post states, old(..) is now allowed in `assert` and
`assume` statements in inlined spec blocks. The primary use case is to
support loop invariants: as showcased in the `verify_vector.move` test.

Note that the old(..) expressions in inline spec block is restricted to
take a single expression that is a function parameter only. No other
cases are currently allowed. Should we want to relax this constraint in
the future, the `ModuleBuilder::check_condition_is_valid` function needs
to be updated and the test case `inline_spec_err.move` in `move-model`
needs to be updated as well.

Closes: #8535