[move-prover] Strengthened some framework specs, added some friend decls, fixed a bug
This takes advantage of the new friends feature and new invariant
scheme that allows temporary disabling of invariants to strengthen
some specifications in the Diem framework.

* Added "friend DiemFramework" to 8 or 9 files.
* Strengthed several invariants in DiemAccount to use <==> instead of where.
* Added several more invariants related to create acount functions
* Removed -v1 option from cargo tests, since it doesn't support friends
* Added friend AccountAdministrationScripts to VASPDomains
* Commented out lines in test cases in
  language/move-lang/functional-tests/tests/diem/
  that relied on new "friend" functions being public. Made some other
  changes on freezing.move.
  (Useful tests will be made into unit tests sometime soon.)

This PR also fixes a bug in the new invariant scheme where "update
invariants" were being assumed and asserted when they shouldn't be,
leading to boogie errors.

Did "cargo run" in the diem-framework directory to re-do docs and mv files.

Closes: #8668