[move-prover] make the no_opaque test ignore internal functions only
Prior to this commit, the no_opaque test ignores `pragma opaque` in
almost every function. Although the prover does not time out in 40s,
this is still a non-trivial amount of time spend in unit test. More
importantly, this might not align with what we have envisioned for how
the specs are supposed to be written: specs are expected to be written
and complete for public-facing functions --- as a result, the
public-facing functions can be opaque from a verification point of view,
hence increasing compositionality of the specs. Inlining public
functions destroy such compositionality.

Closes: #8882
Closes: #10569