[move] Drop `Qualified` in Move IR `ModuleIdent`
> Drop the `Qualified`. Just `ModuleIdent`... it's cleaner.

The commit's parent removed the enum `ModuleIdent`, which had one of two
forms: either a `QualifiedModuleIdent` -- an address and module name --
or a `Transaction` -- the identifier `Transaction` followed by a module
name -- and removed the `Transaction` syntax altogether.

As a result, there's no longer a need to distinguish between a
`ModuleIdent` and a `QualifiedModuleIdent`. So, this commit renames
`QualifiedModuleIdent` to simply `ModuleIdent`, and renames functions
that operate on these identifiers to match.

Closes: #9028