[move-prover] Move generation of well-formed assumptions into its own processor.
This moves the generation of entrypoint assumptions out of spec instrumentation into its own processor. The processor runs *after* global invariant instrumentation, so it sees all memory including that which is only used by injected invariants.

This fixes a bug unearthed during the development of specification variables. Namely, if a global invariant uses memory which is not used by the function in which it is injected, we missed to generate the right assumptions for this memory, because usage analysis happens before invariants are injected. In the new model, we run usage analysis again after all specs have been injected, in the new pipeline step `EntryPointInstrumentationProcessor`.

Updates in baselines are caused by slightly different source locations resulting from the new processor and change in order of processing.

Closes: #9215