[storage] StateDB migration towards supporting get_account_count.
A new storage config (`account_count_migration`) is put in place, upon which the state db starts to write down JellifishMerkleTree internal nodes in a new format in which leaf counts on children will be recorded if available. Once enabled, the DB is incompatible with older versions of the node. Once all accounts get updated at least once, the leaf count should be accessible on the root node, which indicates the completion of the migration.

Closes: #9304