[move-stdlib] use general-purpose GUID's for event handle ID's
- Move general-purpose GUID module from the experimental code to the stdlib
- Add unit tests for GUID
- Use a GUID as the unique ID for an event handle instead of an event-specific id.
- Eliminate all references to `Event::publish_generator`, which is no longer needed.
`GUID::create` lazily publishes a `Generator` under the signer's address if it needs to.

The event/GUID correspondence has some potentially useful applications for objects with
globally unique ID's. For example: say you have NFT's and want to support an efficient API
for looking up an NFT by its ID. You can allow each NFT to carry an event stream whose GUID
is used as the unique ID for the NFT. Transfers of the NFT emit events to this stream. This
lets you query the event stream to find both the current account that holds the NFT (it's
the last event emitted to the stream) and the ownership history of the NFT (each event in
the stream represents a transfer).

Closes: #9395