[move] Document VS Code extension releases
Add documentation for the current release process for new move-analyzer
VS Code extension releases.

Note that the current process is still very manual, and thus error-prone.
I think that, ideally, only collectively vetted updates could be
released (GitHub actions may be one way to achieve this). Unfortunately,
for now, anyone who is a member of the `move` publisher team can create a
release, and a released package is not necessarily guaranteed to match what
is in this GitHub repository. Still, since that is the status quo, I thought
I would at least document it, so that anyone on the team could perform a
release as necessary.

In addition, this commit adds an NPM script "publish," which builds and
runs the tests before publishing the extension.

Closes: #9414