[move-prover] Fixed undeclared $choice function
Under some circumstances involve generic parameters and choice functions,
there can be $choice functions that are called but not declared in Boogie.

The problem is that in spec_translator.rs::translate_spec_funs, there
is a loop that calls translate_spec_fun for each spec function.
Normally, "&self" for both of these functions is a SpecTranslate
struct which includes a field for lifted_choice_infos, a refcell of a
BTreeMap. This is modified in translate_choice_functions.

However, in the loop in translate_spec_funs, there are some cases
where a new SpecTranslator struct is created which sets .type_info to
a new value and clones the remaining fields of "self", then calls
translate_spec_fun with this new struct as its
"self". lifted_choice_infos can be updated during that call, but, after
translate_spec_fun returns, the new SpecTranslator with the updated
lift_choice_infos is lost, reverting back to the previous value.
Unfortunately, that value needs to be preserved for
boogie_translator::translate, which calls of translate_spec_funs, and
emits the $choice functions at the end (in "finalize") after all
spec funs have been processed.

At Meng's suggestion, I just changed lifted_choice_infos to and
Rc<RefCell>> so it's shared everywhere.  That seems to work.

Closes: #9504