[move-prover] warn if unable to instantiate a generic global invariant
We already have a warning if a global invariant is not applicable to any
function if the memory mentioned by the invariant has no overlap with
the memory touched by the function.

However, this is not enough. As can be seen in a simple test case,
```move
module 0x42::Test {
    struct S1 has key, store {}
    struct S2<T: store> has key, store { t: T }

    fun foo(account: signer) {
        move_to(&account, S1 {});
    }
    spec module {
	// invariant I
        invariant<T> exists<S1>(@0x42) ==> exists<S2<T>>(@0x42);
    }
}
```
Invariant `I` is not applicable to this function because we cannot
instantiate the type parameter `T` in this generic invariant when
instrumenting it to function `foo`.

This can be further splitted:

- We cannot instrument it for the entrypoint assumption.
  This is currently allowed (because not injecting an assumptiong
  won't affect the soundness and correctness)

- We cannot instrument it for the checking
  This is not allowed in this commit and will trigger a warning.