[move-model] allow extra checks in type instantiations for built-ins
Some built-in spec functions have more restricted rules on what can go
into a type instantiation than normal user functions.

For example, `exists<T>`, in this case, the instantiation for `T` must
be a struct type and cannot be other types.

This commit added the check for `exists` and `global`.

Before this commit, if we have an invariant
```
invariant<T> exists<T>(@0x2);
```
The invariant will lead to a panic in the model building phase.

With this commit, it triggers an error with some illustration printed
out.