[move-prover] disable #9545 temporarily and catch warnings in testsuite
I mistakenly assumed that warnings are captured in the move prover tests
but turns out that this is not. This commit adds the capturing of
warnings as well.

Consequently, #9545 actually creates many warnings when trying to verify
DPN with Move Prover because a couple of invariants cannot be
instantiated. Although the DPN still veriies, seeing a log of warnings
are not pretty. So, let's temporarily disable the effects of #9545 (and
hence the warnings). This will allow me to investigate a bit more on how
we handle un-related type parameters in a generic invariant.

Closes: #9569