[move] Provide code completion via language server
Implement a completion provider in the move-analyzer language server,
so users of the Visual Studio Code extension see code completion suggestions
as they type.

This first iteration of completion in the language server is as simple as
possible, and performs no semantic analysis. However, it must still lex the
source program. That is because, as soon as move-analyzer informs VS Code that
it provides completions, VS Code stops its internal completion engine entirely.
Originally, the intent of this change to move-analyzer was to *supplement*
VS Code's completion items (for example, if you typed `foo` anywhere in the
source file, VS Code would surface `foo` as a completion item everywhere else)
with Move's keywords and built-in operators. Since there is no easy way to
add to VS Code's simple textual completion engine, this commit implements such
an engine on its own, and adds to that the Move keywords and operators.

Doing so resulted in several decisions and future work items:

* Users spend most of their time editing source code that is not saved to disk.
  They only save that code to disk at certain times, such as when they feel it
  is worth saving. To provide completions, move-analyzer cannot just rely on
  what's on disk, and so must receive notifications from the language server
  client as to what source buffers the user is editing. (This could be improved
  by processing incremental updates from the client.)
* Language servers can define "trigger characters" that prompt the client to
  present completions. For Move and many other languages, `.` and `::` make
  sense for these (for example, `foo.` and `foo::`). Move keywords (`if`,
  `fun`, etc.) and operators (`move` and `copy`) can't appear in these
  positions, so they are filtered out when the user's cursor appears in
  these positions. (This could be improved by performing semantic analysis
  and filtering completion items based on whether the cursor is within a
  `address`, `module`, or `fun` block, based on what identifiers are in
  scope, etc.)
* Currently, the source file being edited is lexed whenever completions are
  requested for that file. Instead, move-analyzer could maintain an
  ever-ready syntax tree view of the entire package being edited, that is
  updated as files are modified. When completions are suggested, they ought
  to be readily available based on this tree, instead of being computed
  reactively as requested.
* I noticed that certain contextual keywords were not included in the
  regex-based TextMate grammar for Move, so I added them. (Once the
  language server begins relying on parsed, semantic analyzed views of
  Move programs, it can begin providing semantically-aware syntax
  highlighting that ensures these are not colored as keywords unless
  appropriate.)

Closes: #9589