[move-model] do model-level simplification in move-model
Previously the simplification runs in spec-flatten, which is very
cumbersome in terms of chaining the simplification passes (due to the
fact that the `GlobalEnv` is not clonable).

This commit relotates the transformation on the `GlobalEnv` side into
the move-model, i.e., the simplification transformations will be
co-located with the routines that build the `GlobalEnv`.

This also adds CLI support for invoking a single simplification pass
(right now the `--simplify inline` pass only) with the potential of
constructing a pipeline of passes by applying `--simplify` multiple
times.

The simplified spec should still prove. This is point checked by
```bash
mvp diem-move/diem-framework/core/sources/AccountCreationScripts.move \
    -d diem-move/diem-framework/core/sources \
    -d language/move-stdlib/sources \
    --simplify inline
```

We might need a more thorough check for it, which will come in
the next commit.