[move] Rename `compiler` crate `move-ir-compiler`
There was once a time where the only compiled language in the project
was Move IR, and there was only a single compiler for it, so the name
"compiler" was unambiguous. However, now that the Move source language
exists, it's clearer to rename the `compiler` crate to
`move-ir-compiler`.

I feel now is a good time to do this, because I'd like to add a
transactional test runner crate for the Move IR compiler, and that
involves giving it a name. I could name it
`compiler-transactional-tests`, but that would create yet another
ambiguously named crate in the repository. I figure it's better to
rename the `compiler` crate first, then add the test runner.

Closes: #9782