[move] Add test option to print script bytecode
Currently the transactional test runner's "print-bytecode" command
assumes that the input is a module. This was originally because the
disassembler did not support printing a script's bytecode. However, in
preparation for adding that support to the disassembler, add an option
to have the print-bytecode command treat the input string as a Move IR
module or a script.