[move-prover] Loop invariants for functional macros + fixes
This contains a new test case in [`macro_verification.move`]() which simulates expansion of planned functional macros. The purpose of this to see what we can verify, and what part of the loop invariant can be derived from the specific macro.

Interestingly, after some tweaking, we are able to verify both `foreach!` style and `reduce!` style macros. The later is particularly tricky as it cannot be expressed by a quantifier. Rather, a recursive function is needed to define the loop invariant. This trick has been copied from Dafny, and it seems to work well (however, not with cvc5).

The PR fixes also a few other things around invariants:

- enabling recursive specification functions (this was not possible before)
- supporting a more adequate syntax for loop invariants: `while (cond) body spec` is syntactic sugar for `while ({spec; cond}) body`
- enabling tracing of assertions stemming from loop invariants and other in-code specs.

Also, the prover tests had been broken (perhaps invisible on CI) by a change in the package system, and unfortunately now need to be executed sequentially.

Closes: #9978